requirements.txt
langchain
tiktoken
faiss-cpu
langchain_experimental
python-dotenv
openai